Abbas's Assignment
I apologize for the delay in sending the files due to an unexpected urgent task. The task is still ongoing and I'm working diligently on it.

Anyway, The code isn't fully complete yet, but it provides a glimpse of my coding style. I'm open to adjusting it based on your feedback.